requirements.txt
twarc-csv==0.6.0
twarc
pandas==1.5.1
python-dotenv==0.19.0
plotly-express==0.4.1